refactor(library/Attempt): moves contents of "/utilities" to "/tryCatchStatements"
