+++ src/library/Attempt/Adapted/to.ts
@@ -13,7 +13,7 @@ import type {
 
 import {
   safe,
-} from '../utilities/safeTryCatch';
+} from '../tryCatchStatements/safeTryCatch';
 
 import type Attempt_Adapted_Config from './Config';
