+++ src/library/Attempt/tryCatchStatements/safeTryCatch.ts
